FIXED: Do not call shutdown() on sockets in Windows.
The previous implementation called shutdown() on the sending channal
in e.g., http_open/3.  This causes some servers to abort the entire
transaction.  The Windows socket API went through many iterations, so
I'm not sure why the shutdown is there in the first place.  Currently
the code is more uniform between Unix and Windows and -according to
Matt Lilley-, the change does not cause leaking of memory or handles;
the reason for many of the iterations in the implementation.