FIXED: Windows: possibly deadlock in socket write.  Keri Harris.
If a write-requests is queued and the socket thread receives an
FD_CLOSE without error for the socket, this is not translated in
an error and the socket will wait forever for an FD_WRITE event.

Now, an FD_CLOSE causes the write to return.  If nothing is written
yet, with an error.  Otherwise the amount written is returned.  If
the client wishes to continue writing, it will encounter a closed
socket and raise an error.