ADDED: extra_streams() option to process_create/3
This enables the Unix-style pattern of passing additional file
descriptors to child processes beyond stdin, stdout, and stderr, so that
parent and child can communicate on multiple streams. For example, a
Prolog program could launch a fully-interactive child process, passing
stdin/stdout/stderr connected to the TTY, while reserving a separate
pair of descriptors for programmatic communication with the child.